Update builder lib to 1.0.0
