+++ pom.xml
@@ -41,8 +41,8 @@
 
     <!-- Attention - make sure that the tools version is the versions of 
       the transitive dependencies of the builder version -->
-    <android.builder.version>0.13.3</android.builder.version>
-    <android.tools.version>23.1.3</android.tools.version>
+    <android.builder.version>1.0.0</android.builder.version>
+    <android.tools.version>24.0.0</android.tools.version>
   </properties>
 
   <scm>
